backend/requirements.txt
django>=5.1,<5.2
django-cors-headers>=4.6,<4.7
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-filter>=24.3,<24.4
